Score Counter and AppleDestroyed
Created a score counter for the game that adds 100 points to the player's score in the top right corner of the game every time the player picks up an apple with the basket. Also added a way for the game to know when a player misses an apple and destroys the apples when a player misses one.